+++ src/commands/EntityCreateCommand.ts
@@ -1,5 +1,5 @@
-import * as fs from "fs";
-import {ConnectionOptions} from "../connection/ConnectionOptions";
+import {ConnectionOptionsReader} from "../connection/ConnectionOptionsReader";
+import {CommandUtils} from "./CommandUtils";
 
 /**
  * Generates a new entity.
@@ -26,7 +26,7 @@ export class EntityCreateCommand {
             })
             .option("cf", {
                 alias: "config",
-                default: "ormconfig.json",
+                default: "ormconfig",
                 describe: "Name of the file with connection configuration."
             });
     }
@@ -39,33 +39,18 @@ export class EntityCreateCommand {
         // if directory is not set then try to open tsconfig and find default path there
         if (!directory) {
             try {
-                const connections: ConnectionOptions[] = require(process.cwd() + "/" + argv.config);
-                if (connections) {
-                    const connection = connections.find(connection => { // todo: need to implement "environment" support in the ormconfig too
-                        return connection.name === argv.connection || ((argv.connection === "default" || !argv.connection) && !connection.name);
-                    });
-                    if (connection && connection.cli) {
-                        directory = connection.cli.entitiesDir;
-                    }
-                }
+                const connectionOptionsReader = new ConnectionOptionsReader({ root: process.cwd(), configName: argv.config });
+                const connectionOptions = await connectionOptionsReader.get(argv.connection);
+                directory = connectionOptions.cli ? connectionOptions.cli.entitiesDir : undefined;
             } catch (err) { }
         }
 
-        await EntityCreateCommand.createFile(process.cwd() + "/" + (directory ? (directory + "/") : "") + filename, fileContent);
+        await CommandUtils.createFile(process.cwd() + "/" + (directory ? (directory + "/") : "") + filename, fileContent);
     }
 
     // -------------------------------------------------------------------------
     // Protected Static Methods
     // -------------------------------------------------------------------------
-
-    /**
-     * Creates a file with the given content in the given path.
-     */
-    protected static createFile(path: string, content: string): Promise<void> {
-        return new Promise<void>((ok, fail) => {
-            fs.writeFile(path, content, err => err ? fail(err) : ok());
-        });
-    }
 
     /**
      * Gets contents of the entity file.