+++ src/connection/Connection.ts
@@ -5,21 +5,18 @@ import {RepositoryNotFoundError} from "./error/RepositoryNotFoundError";
 import {ObjectType} from "../common/ObjectType";
 import {EntityManager} from "../entity-manager/EntityManager";
 import {DefaultNamingStrategy} from "../naming-strategy/DefaultNamingStrategy";
-import {CannotCloseNotConnectedError} from "./error/CannotCloseNotConnectedError";
+import {CannotExecuteNotConnectedError} from "./error/CannotExecuteNotConnectedError";
 import {CannotConnectAlreadyConnectedError} from "./error/CannotConnectAlreadyConnectedError";
 import {TreeRepository} from "../repository/TreeRepository";
 import {NamingStrategyInterface} from "../naming-strategy/NamingStrategyInterface";
 import {RepositoryNotTreeError} from "./error/RepositoryNotTreeError";
-import {CannotSyncNotConnectedError} from "./error/CannotSyncNotConnectedError";
 import {SpecificRepository} from "../repository/SpecificRepository";
 import {EntityMetadata} from "../metadata/EntityMetadata";
-import {SchemaBuilder} from "../schema-builder/SchemaBuilder";
 import {Logger} from "../logger/Logger";
 import {QueryRunnerProvider} from "../query-runner/QueryRunnerProvider";
 import {EntityMetadataNotFound} from "../metadata-args/error/EntityMetadataNotFound";
 import {MigrationInterface} from "../migration/MigrationInterface";
 import {MigrationExecutor} from "../migration/MigrationExecutor";
-import {CannotRunMigrationNotConnectedError} from "./error/CannotRunMigrationNotConnectedError";
 import {PlatformTools} from "../platform/PlatformTools";
 import {MongoRepository} from "../repository/MongoRepository";
 import {MongoDriver} from "../driver/mongodb/MongoDriver";
@@ -42,7 +39,7 @@ import {ConnectionMetadataBuilder} from "./ConnectionMetadataBuilder";
 export class Connection {
 
     // -------------------------------------------------------------------------
-    // Public Readonly properties
+    // Public Readonly Properties
     // -------------------------------------------------------------------------
 
     /**
@@ -145,9 +142,22 @@ export class Connection {
         // set connected status for the current connection
         Object.assign(this, { isConnected: true });
 
-        // build all metadatas registered in the current connection
         try {
+
+            // build all metadatas registered in the current connection
             this.buildMetadatas();
+
+            // if option is set - drop schema once connection is done
+            if (this.options.dropSchemaOnConnection && !PlatformTools.getEnvVariable("SKIP_SCHEMA_CREATION"))
+                await this.dropDatabase();
+
+            // if option is set - automatically synchronize a schema
+            if (this.options.autoSchemaSync && !PlatformTools.getEnvVariable("SKIP_SCHEMA_CREATION"))
+                await this.syncSchema();
+
+            // if option is set - automatically synchronize a schema
+            if (this.options.autoMigrationsRun && !PlatformTools.getEnvVariable("SKIP_MIGRATIONS_RUN"))
+                await this.runMigrations();
 
         } catch (error) {
 
@@ -166,7 +176,7 @@ export class Connection {
      */
     async close(): Promise<void> {
         if (!this.isConnected)
-            throw new CannotCloseNotConnectedError(this.name);
+            throw new CannotExecuteNotConnectedError(this.name);
 
         await this.driver.disconnect();
         Object.assign(this, { isConnected: false });
@@ -181,23 +191,17 @@ export class Connection {
     async syncSchema(dropBeforeSync: boolean = false): Promise<void> {
 
         if (!this.isConnected)
-            throw new CannotCloseNotConnectedError(this.name);
+            throw new CannotExecuteNotConnectedError(this.name);
 
         if (dropBeforeSync)
             await this.dropDatabase();
 
-        if (this.driver instanceof MongoDriver) { // todo: temporary
-            await this.driver.syncSchema(this.entityMetadatas);
-
-        } else {
-            const schemaBuilder = new SchemaBuilder(this.driver, this.logger, this.entityMetadatas);
-            await schemaBuilder.build();
-        }
+        await this.driver.syncSchema();
     }
 
     /**
      * Drops the database and all its data.
-     * Be careful with this method on production since this method will erase all your database tables and data inside them.
+     * Be careful with this method on production since this method will erase all your database tables and their data.
      * Can be used only after connection to the database is established.
      */
     async dropDatabase(): Promise<void> {
@@ -212,7 +216,7 @@ export class Connection {
     async runMigrations(): Promise<void> {
 
         if (!this.isConnected)
-            throw new CannotCloseNotConnectedError(this.name);
+            throw new CannotExecuteNotConnectedError(this.name);
 
         const migrationExecutor = new MigrationExecutor(this);
         await migrationExecutor.executePendingMigrations();
@@ -225,7 +229,7 @@ export class Connection {
     async undoLastMigration(): Promise<void> {
 
         if (!this.isConnected)
-            throw new CannotCloseNotConnectedError(this.name);
+            throw new CannotExecuteNotConnectedError(this.name);
 
         const migrationExecutor = new MigrationExecutor(this);
         await migrationExecutor.undoLastMigration();
@@ -239,7 +243,7 @@ export class Connection {
     }
 
     /**
-     Gets entity metadata for the given entity class or schema name.
+     * Gets entity metadata for the given entity class or schema name.
      */
     getMetadata(target: Function|string): EntityMetadata {
         const metadata = this.findMetadata(target);
@@ -276,6 +280,7 @@ export class Connection {
 
     /**
      * Gets mongodb-specific repository for the given entity class or name.
+     * Works only if connection is mongodb-specific.
      */
     getMongoRepository<Entity>(target: ObjectType<Entity>|string): MongoRepository<Entity> {
         if (!(this.driver instanceof MongoDriver))
@@ -295,7 +300,7 @@ export class Connection {
     }
 
     /**
-     * Wraps given function execution (and all operations made there) in a transaction.
+     * Wraps given function execution (and all operations made there) into a transaction.
      * All database operations must be executed using provided entity manager.
      */
     async transaction(runInTransaction: (entityManger: EntityManager) => Promise<any>,
@@ -308,7 +313,7 @@ export class Connection {
 
         const usedQueryRunnerProvider = queryRunnerProvider || new QueryRunnerProvider(this.driver, true);
         const queryRunner = await usedQueryRunnerProvider.provide();
-        const transactionEntityManager = new EntityManager(this, usedQueryRunnerProvider);
+        const transactionEntityManager = new EntityManagerFactory().create(this, usedQueryRunnerProvider);
 
         try {
             await queryRunner.beginTransaction();
@@ -379,7 +384,7 @@ export class Connection {
     /**
      * Creates a new entity manager with a single opened connection to the database.
      * This may be useful if you want to perform all db queries within one connection.
-     * After finishing with entity manager, don't forget to release it, to release connection back to pool.
+     * After finishing with entity manager, don't forget to release it (to release database connection back to pool).
      */
     createIsolatedManager(queryRunnerProvider?: QueryRunnerProvider): EntityManager {
         if (!queryRunnerProvider)
@@ -391,7 +396,7 @@ export class Connection {
     /**
      * Creates a new repository with a single opened connection to the database.
      * This may be useful if you want to perform all db queries within one connection.
-     * After finishing with entity manager, don't forget to release it, to release connection back to pool.
+     * After finishing with entity manager, don't forget to release it (to release database connection back to pool).
      */
     createIsolatedRepository<Entity>(entityClassOrName: ObjectType<Entity>|string, queryRunnerProvider?: QueryRunnerProvider): Repository<Entity> {
         if (!queryRunnerProvider)
@@ -403,7 +408,7 @@ export class Connection {
     /**
      * Creates a new specific repository with a single opened connection to the database.
      * This may be useful if you want to perform all db queries within one connection.
-     * After finishing with entity manager, don't forget to release it, to release connection back to pool.
+     * After finishing with entity manager, don't forget to release it (to release database connection back to pool).
      */
     createIsolatedSpecificRepository<Entity>(entityClassOrName: ObjectType<Entity>|string, queryRunnerProvider?: QueryRunnerProvider): SpecificRepository<Entity> {
         if (!queryRunnerProvider)
@@ -443,7 +448,7 @@ export class Connection {
     // -------------------------------------------------------------------------
 
     /**
-     * Finds entity metadata exist for the given entity class, target name or table name.
+     * Finds exist entity metadata by the given entity class, target name or table name.
      */
     protected findMetadata(target: Function|string): EntityMetadata|undefined {
         return this.entityMetadatas.find(metadata => {
@@ -457,7 +462,7 @@ export class Connection {
     }
 
     /**
-     * Builds all registered metadatas.
+     * Builds metadatas for all registered classes inside this connection.
      */
     protected buildMetadatas(): void {
 
@@ -465,7 +470,7 @@ export class Connection {
         const repositoryFactory = new RepositoryFactory();
         const entityMetadataValidator = new EntityMetadataValidator();
 
-        // build subscribers if they are not disallowed from high-level (for example they can disallowed from migrations run process)
+        // create subscribers instances if they are not disallowed from high-level (for example they can disallowed from migrations run process)
         if (!PlatformTools.getEnvVariable("SKIP_SUBSCRIBERS_LOADING")) {
             const subscribers = connectionMetadataBuilder.buildSubscribers(this.options.subscribers || []);
             Object.assign(this, { subscribers: subscribers });